fastapi/requirements.txt
fastapi
pydantic
uvicorn
openai
pymupdf
slack-sdk
python-dotenv
gunicorn
